+++ src/view/app.py
@@ -490,7 +490,7 @@ class App(ViewApp):
 
     def context(self, r_or_none: RouteOrCallable | None = None):
         return context_impl(r_or_none)
-
+    
     async def _app(self, scope, receive, send) -> None:
         return await self.asgi_app_entry(scope, receive, send)